backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Aditya Banerjee <aditya_banerjee@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
llamaapi = "^0.1.36"
python-dotenv = "^1.0.1"

[tool.poetry.scripts]
start = "poetry run python backend/main.py"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
